Extract trust verification into shared runt-trust crate
Creates a new runt-trust crate that can be used by both the notebook app
and the daemon for trust verification. This is a pure refactor with no
behavior changes.

The notebook crate now re-exports from runt-trust, maintaining API
compatibility for existing code.
