+++ crates/runt-trust/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "runt-trust"
+version = "0.1.0"
+edition = "2021"
+description = "Notebook trust verification using HMAC signatures"
+
+[dependencies]
+dirs = "5.0"
+hex = "0.4"
+hmac = "0.12"
+rand = "0.8"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+sha2 = "0.10"
+
+[dev-dependencies]
+serial_test = "3.0"
+tempfile = "3.10"
